# swagger.yaml
openapi: 3.0.0
info:
  title: The System MySQL Database
  description: The System MySQL Database
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
...
  /accounts:
    get:
      summary: Get a list of all accounts
      description: Retrieve list of all accounts.
      operationId: getAllAccounts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: An array of all accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 123
                    accountNumber:
                      type: string
                      example: 'AAA001'
                    accountName:
                      type: string
                      example: 'Disney'
                    openDate:
                      type: string
                      minLength: 8
                      maxLength: 8
                      format: date
                      pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                      example: 20190816
                      description: 'Date the account was opened.'
                    debtorAge:
                      type: integer
                      example: 30
                    paymentTermDays:
                      type: integer
                      example: 60
                    creditLimit:
                      type: integer
                      example: 1500.00
                    totalBalance:
                      type: integer
                      example: 1000.00
                    amountDue:
                      type: integer
                      example: 1500.00
                    currentBalance:
                      type: integer
                      example: 1000.00
                    days30:
                      type: integer
                      example: 30.00
                    days60:
                      type: integer
                      example: 60.00
                    days90:
                      type: integer
                      example: 90.00
                    days120:
                      type: integer
                      example: 120.00
                    days150:
                      type: integer
                      example: 150.00
                    days180:
                      type: integer
                      example: 180.00
                    days180Over:
                      type: integer
                      example: 200.00
                    paymentMethod:
                      type: string
                      example: 'EFT'
                    paymentDueDate:
                      type: integer
                      example: 2
                    debitOrderDate:
                      type: integer
                      example: 2
                    lastPaymentDate:
                      type: string
                      minLength: 8
                      maxLength: 8
                      format: date
                      pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                      example: 20190816
                    lastPaymentAmount:
                      type: integer
                      example: 2016.00
                    lastPTPDate:
                      type: string
                      minLength: 8
                      maxLength: 8
                      format: date
                      pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                      example: 20190816
                    lastPTPAmount:
                      type: integer
                      example: 1016.00
                    accountNotes:
                      type: string
                      example: 'Customer can be shouty'
                    accountStatus:
                      type: string
                      example: 'Open'
                    arg:
                      type: string
                      example: 'A1'
                    createdDate:
                      type: string
                      example: '2020-05-05T09:12:57.848Z'
                    createdBy:
                      type: string
                      example: 'mickey@example.com'
                    updatedDate:
                      type: string
                      example: '2020-05-08T03:11:21.553Z'
                    updatedBy:
                      type: string
                      example: 'mickey@example.com'
                    f_customerId:
                      type: integer
                      example: 123
        '401':
          $ref: '#/components/responses/UnauthorisedError'
    post:
      summary: Create a new account
      description: Restricted to admin users.
      operationId: createAccount
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  accountNumber:
                    type: string
                    example: 'AAA001'
                  accountName:
                    type: string
                    example: 'Disney'
                  openDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                    description: 'Date the account was opened.'
                  debtorAge:
                    type: integer
                    example: 30
                  paymentTermDays:
                    type: integer
                    example: 60
                  creditLimit:
                    type: integer
                    example: 1500.00
                  totalBalance:
                    type: integer
                    example: 1000.00
                  amountDue:
                    type: integer
                    example: 1500.00
                  currentBalance:
                    type: integer
                    example: 1000.00
                  days30:
                    type: integer
                    example: 30.00
                  days60:
                    type: integer
                    example: 60.00
                  days90:
                    type: integer
                    example: 90.00
                  days120:
                    type: integer
                    example: 120.00
                  days150:
                    type: integer
                    example: 150.00
                  days180:
                    type: integer
                    example: 180.00
                  days180Over:
                    type: integer
                    example: 200.00
                  paymentMethod:
                    type: string
                    example: 'EFT'
                  paymentDueDate:
                    type: integer
                    example: 2
                  debitOrderDate:
                    type: integer
                    example: 2
                  lastPaymentDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                  lastPaymentAmount:
                    type: integer
                    example: 2016.00
                  lastPTPDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                  lastPTPAmount:
                    type: integer
                    example: 1016.00
                  accountNotes:
                    type: string
                    example: 'Customer can be shouty'
                  accountStatus:
                    type: string
                    example: 'Open'
                  arg:
                    type: string
                    example: 'A1'
                  createdDate:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  createdBy:
                    type: string
                    example: 'mickey@example.com'
                  updatedDate:
                    type: string
                    example: '2020-05-08T03:11:21.553Z'
                  updatedBy:
                    type: string
                    example: 'mickey@example.com'
                  f_customerId:
                    type: string
                    example: 'AAA101'
              required:
                - accountNumber
                - updatedDate
                - updatedBy
      responses:
        '200':
          description: Account created successfully. The account number of the new account is returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountNumber:
                    type: string
                    example: 'AAA001'
        '400':
          description: Account name is already on the books
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Account name 'Disney' is already on the books"
        '401':
          $ref: '#/components/responses/UnauthorisedError'
  /accounts/{accountNumber}:
    parameters:
      - in: path
        name: accountNumber
        description: Account number
        required: true
        example: 'AAA001'
        schema:
          type: string
    get:
      summary: Get a single account by account number
      description: Retrieve a single account by account number.
      operationId: getAccountByAccountNumber
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Details of the specified account
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  accountNumber:
                    type: string
                    example: 'AAA001'
                  accountName:
                    type: string
                    example: 'Disney'
                  openDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                    description: 'Date the account was opened.'
                  debtorAge:
                    type: integer
                    example: 30
                  paymentTermDays:
                    type: integer
                    example: 60
                  creditLimit:
                    type: integer
                    example: 1500.00
                  totalBalance:
                    type: integer
                    example: 1000.00
                  amountDue:
                    type: integer
                    example: 1500.00
                  currentBalance:
                    type: integer
                    example: 1000.00
                  days30:
                    type: integer
                    example: 30.00
                  days60:
                    type: integer
                    example: 60.00
                  days90:
                    type: integer
                    example: 90.00
                  days120:
                    type: integer
                    example: 120.00
                  days150:
                    type: integer
                    example: 150.00
                  days180:
                    type: integer
                    example: 180.00
                  days180Over:
                    type: integer
                    example: 200.00
                  paymentMethod:
                    type: string
                    example: 'EFT'
                  paymentDueDate:
                    type: integer
                    example: 2
                  debitOrderDate:
                    type: integer
                    example: 2
                  lastPaymentDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                  lastPaymentAmount:
                    type: integer
                    example: 2016.00
                  lastPTPDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                  lastPTPAmount:
                    type: integer
                    example: 1016.00
                  accountNotes:
                    type: string
                    example: 'Customer can be shouty'
                  accountStatus:
                    type: string
                    example: 'Open'
                  arg:
                    type: string
                    example: 'A1'
                  createdDate:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  createdBy:
                    type: string
                    example: 'mickey@example.com'
                  updatedDate:
                    type: string
                    example: '2020-05-08T03:11:21.553Z'
                  updatedBy:
                    type: string
                    example: 'mickey@example.com'
                  f_customerId:
                    type: string
                    example: 'AAA101'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorisedError'
    put:
      summary: Update an account
      description: Update details of an account.
      operationId: updateAccount
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 123
                accountNumber:
                  type: string
                  example: 'AAA001'
                accountName:
                  type: string
                  example: 'Disney'
                openDate:
                  type: string
                  minLength: 8
                  maxLength: 8
                  format: date
                  pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                  example: 20190816
                  description: 'Date the account was opened.'
                debtorAge:
                  type: integer
                  example: 30
                paymentTermDays:
                  type: integer
                  example: 60
                creditLimit:
                  type: integer
                  example: 1500.00
                totalBalance:
                  type: integer
                  example: 1000.00
                amountDue:
                  type: integer
                  example: 1500.00
                currentBalance:
                  type: integer
                  example: 1000.00
                days30:
                  type: integer
                  example: 30.00
                days60:
                  type: integer
                  example: 60.00
                days90:
                  type: integer
                  example: 90.00
                days120:
                  type: integer
                  example: 120.00
                days150:
                  type: integer
                  example: 150.00
                days180:
                  type: integer
                  example: 180.00
                days180Over:
                  type: integer
                  example: 200.00
                paymentMethod:
                  type: string
                  example: 'EFT'
                paymentDueDate:
                  type: integer
                  example: 2
                debitOrderDate:
                  type: integer
                  example: 2
                lastPaymentDate:
                  type: string
                  minLength: 8
                  maxLength: 8
                  format: date
                  pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                  example: 20190816
                lastPaymentAmount:
                  type: integer
                  example: 2016.00
                lastPTPDate:
                  type: string
                  minLength: 8
                  maxLength: 8
                  format: date
                  pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                  example: 20190816
                lastPTPAmount:
                  type: integer
                  example: 1016.00
                accountNotes:
                  type: string
                  example: 'Customer can be shouty'
                accountStatus:
                  type: string
                  example: 'Open'
                arg:
                  type: string
                  example: 'A1'
                createdDate:
                  type: string
                  example: '2020-05-05T09:12:57.848Z'
                createdBy:
                  type: string
                  example: 'mickey@example.com'
                updatedDate:
                  type: string
                  example: '2020-05-08T03:11:21.553Z'
                updatedBy:
                  type: string
                  example: 'mickey@example.com'
                f_customerId:
                  type: string
                  example: 'AAA101'
      responses:
        '200':
          description: Account updated successfully. The details of the updated account are returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  accountNumber:
                    type: string
                    example: 'AAA001'
                  accountName:
                    type: string
                    example: 'Disney'
                  openDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                    description: 'Date the account was opened.'
                  debtorAge:
                    type: integer
                    example: 30
                  paymentTermDays:
                    type: integer
                    example: 60
                  creditLimit:
                    type: integer
                    example: 1500.00
                  totalBalance:
                    type: integer
                    example: 1000.00
                  amountDue:
                    type: integer
                    example: 1500.00
                  currentBalance:
                    type: integer
                    example: 1000.00
                  days30:
                    type: integer
                    example: 30.00
                  days60:
                    type: integer
                    example: 60.00
                  days90:
                    type: integer
                    example: 90.00
                  days120:
                    type: integer
                    example: 120.00
                  days150:
                    type: integer
                    example: 150.00
                  days180:
                    type: integer
                    example: 180.00
                  days180Over:
                    type: integer
                    example: 200.00
                  paymentMethod:
                    type: string
                    example: 'EFT'
                  paymentDueDate:
                    type: integer
                    example: 2
                  debitOrderDate:
                    type: integer
                    example: 2
                  lastPaymentDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                  lastPaymentAmount:
                    type: integer
                    example: 2016.00
                  lastPTPDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                  lastPTPAmount:
                    type: integer
                    example: 1016.00
                  accountNotes:
                    type: string
                    example: 'Customer can be shouty'
                  accountStatus:
                    type: string
                    example: 'Open'
                  arg:
                    type: string
                    example: 'A1'
                  createdDate:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  createdBy:
                    type: string
                    example: 'mickey@example.com'
                  updatedDate:
                    type: string
                    example: '2020-05-08T03:11:21.553Z'
                  updatedBy:
                    type: string
                    example: 'mickey@example.com'
                  f_customerId:
                    type: string
                    example: 'AAA101'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorisedError'
    delete:
      summary: Delete an account
      description: Delete an account.
      operationId: deleteAccount
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Account deleted successfully'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorisedError'
...